fix for legend in layerlist widget
